Refactored k8sutil into seperate files
