+++ pkg/k8sutil/k8sutil.go
@@ -43,7 +43,6 @@ import (
 	"k8s.io/client-go/pkg/api/v1"
 	apps "k8s.io/client-go/pkg/apis/apps/v1beta1"
 	"k8s.io/client-go/pkg/apis/extensions/v1beta1"
-	storage "k8s.io/client-go/pkg/apis/storage/v1beta1"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/tools/clientcmd"
@@ -291,260 +290,10 @@ func (k *K8sutil) CreateKubernetesThirdPartyResource() error {
 	return nil
 }
 
-// DeleteServices creates the discovery service
-func (k *K8sutil) DeleteServices(clusterName, namespace string) {
-
-	fullDiscoveryServiceName := fmt.Sprintf("%s-%s", discoveryServiceName, clusterName)
-	err := k.Kclient.CoreV1().Services(namespace).Delete(fullDiscoveryServiceName, &metav1.DeleteOptions{})
-	if err != nil {
-		logrus.Error("Could not delete service "+fullDiscoveryServiceName+":", err)
-	} else {
-		logrus.Infof("Delete service: %s", fullDiscoveryServiceName)
-	}
-
-	fullDataServiceName := dataServiceName + "-" + clusterName
-	err = k.Kclient.CoreV1().Services(namespace).Delete(fullDataServiceName, &metav1.DeleteOptions{})
-	if err != nil {
-		logrus.Error("Could not delete service "+fullDataServiceName+":", err)
-	} else {
-		logrus.Infof("Delete service: %s", fullDataServiceName)
-	}
-
-	fullClientServiceName := clientServiceName + "-" + clusterName
-	err = k.Kclient.CoreV1().Services(namespace).Delete(fullClientServiceName, &metav1.DeleteOptions{})
-	if err != nil {
-		logrus.Error("Could not delete service "+fullClientServiceName+":", err)
-	} else {
-		logrus.Infof("Delete service: %s", fullClientServiceName)
-	}
-
-}
-
-// CreateDiscoveryService creates the discovery service
-func (k *K8sutil) CreateDiscoveryService(clusterName, namespace string) error {
-
-	fullDiscoveryServiceName := fmt.Sprintf("%s-%s", discoveryServiceName, clusterName)
-	component := "elasticsearch" + "-" + clusterName
-	// Check if service exists
-	svc, err := k.Kclient.CoreV1().Services(namespace).Get(fullDiscoveryServiceName, metav1.GetOptions{})
-
-	// Service missing, create
-	if len(svc.Name) == 0 {
-		logrus.Info("Discovery Service not found, creating...")
-
-		discoverySvc := &v1.Service{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: fullDiscoveryServiceName,
-				Labels: map[string]string{
-					"component": component,
-					"role":      "master",
-				},
-			},
-			Spec: v1.ServiceSpec{
-				Selector: map[string]string{
-					"component": component,
-					"role":      "master",
-				},
-				Ports: []v1.ServicePort{
-					v1.ServicePort{
-						Name:     "transport",
-						Port:     9300,
-						Protocol: "TCP",
-					},
-				},
-			},
-		}
-
-		_, err := k.Kclient.CoreV1().Services(namespace).Create(discoverySvc)
-
-		if err != nil {
-			logrus.Error("Could not create discovery service! ", err)
-			return err
-		}
-	} else if err != nil {
-		logrus.Error("Could not get discovery service! ", err)
-		return err
-	}
-
-	return nil
-}
-
-// CreateDataService creates the data service
-func (k *K8sutil) CreateDataService(clusterName, namespace string) error {
-	fullDataServiceName := dataServiceName + "-" + clusterName
-	component := "elasticsearch" + "-" + clusterName
-	// Check if service exists
-	svc, err := k.Kclient.CoreV1().Services(namespace).Get(fullDataServiceName, metav1.GetOptions{})
-
-	// Service missing, create
-	if len(svc.Name) == 0 {
-		logrus.Infof("%s not found, creating...", fullDataServiceName)
-
-		dataService := &v1.Service{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: fullDataServiceName,
-				Labels: map[string]string{
-					"service.alpha.kubernetes.io/tolerate-unready-endpoints": "true",
-				},
-				Annotations: map[string]string{
-					"component": component,
-					"name":      fullDataServiceName,
-				},
-			},
-			Spec: v1.ServiceSpec{
-				Selector: map[string]string{
-					"component": component,
-					"role":      "data",
-				},
-				Ports: []v1.ServicePort{
-					v1.ServicePort{
-						Name:     "transport",
-						Port:     9300,
-						Protocol: "TCP",
-					},
-				},
-			},
-		}
-
-		_, err := k.Kclient.CoreV1().Services(namespace).Create(dataService)
-
-		if err != nil {
-			logrus.Error("Could not create data service", err)
-			return err
-		}
-	} else if err != nil {
-		logrus.Error("Could not get data service! ", err)
-		return err
-	}
-
-	return nil
-}
-
-// GetClientServiceNameFullDNS return the full DNS name of the client service
-func (k *K8sutil) GetClientServiceNameFullDNS(clusterName, namespace string) string {
-	return fmt.Sprintf("%s.%s.svc.cluster.local", k.GetClientServiceName(clusterName), namespace)
-}
-
-// GetClientServiceName return the name of the client service
-func (k *K8sutil) GetClientServiceName(clusterName string) string {
-	return fmt.Sprintf("%s-%s", clientServiceName, clusterName)
-}
-
-// CreateClientService creates the client service
-func (k *K8sutil) CreateClientService(clusterName, namespace string, nodePort int32) error {
-
-	fullClientServiceName := k.GetClientServiceName(clusterName)
-	component := fmt.Sprintf("elasticsearch-%s", clusterName)
-	// Check if service exists
-	svc, err := k.Kclient.CoreV1().Services(namespace).Get(fullClientServiceName, metav1.GetOptions{})
-
-	// Service missing, create
-	if len(svc.Name) == 0 {
-		logrus.Infof("%s not found, creating...", fullClientServiceName)
-
-		clientSvc := &v1.Service{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: fullClientServiceName,
-				Labels: map[string]string{
-					"component": component,
-					"role":      "client",
-				},
-			},
-			Spec: v1.ServiceSpec{
-				Selector: map[string]string{
-					"component": component,
-					"role":      "client",
-				},
-				Ports: []v1.ServicePort{
-					v1.ServicePort{
-						Name:     "http",
-						Port:     9200,
-						Protocol: "TCP",
-					},
-				},
-			},
-		}
-		if nodePort > 0 {
-			clientSvc.Spec.Type = v1.ServiceTypeNodePort
-			clientSvc.Spec.Ports[0].NodePort = nodePort
-		}
-
-		_, err := k.Kclient.CoreV1().Services(namespace).Create(clientSvc)
-
-		if err != nil {
-			logrus.Error("Could not create client service", err)
-			return err
-		}
-	} else if err != nil {
-		logrus.Error("Could not get client service! ", err)
-		return err
-	}
-
-	return nil
-}
-
-// DeleteClientMasterDeployment deletes the client or master deployment
-func (k *K8sutil) DeleteClientMasterDeployment(deploymentType, clusterName, namespace string) error {
-
-	labelSelector := ""
-
-	if deploymentType == "client" {
-		labelSelector = "component=elasticsearch" + "-" + clusterName + ",role=client"
-	} else if deploymentType == "master" {
-		labelSelector = "component=elasticsearch" + "-" + clusterName + ",role=master"
-	}
-
-	// Get list of deployments
-	deployments, err := k.Kclient.ExtensionsV1beta1().Deployments(namespace).List(metav1.ListOptions{LabelSelector: labelSelector})
-
-	if err != nil {
-		logrus.Error("Could not get deployments! ", err)
-	}
-
-	for _, deployment := range deployments.Items {
-		//Scale the deployment down to zero (https://github.com/kubernetes/client-go/issues/91)
-		deployment.Spec.Replicas = new(int32)
-		deployment, err := k.Kclient.ExtensionsV1beta1().Deployments(namespace).Update(&deployment)
-
-		if err != nil {
-			logrus.Errorf("Could not scale deployment: %s ", deployment.Name)
-		} else {
-			logrus.Infof("Scaled deployment: %s to zero", deployment.Name)
-		}
-
-		err = k.Kclient.ExtensionsV1beta1().Deployments(namespace).Delete(deployment.Name, &metav1.DeleteOptions{})
-
-		if err != nil {
-			logrus.Errorf("Could not delete deployments: %s ", deployment.Name)
-		} else {
-			logrus.Infof("Deleted deployment: %s", deployment.Name)
-		}
-	}
-
-	// Get list of ReplicaSets
-	replicaSets, err := k.Kclient.ExtensionsV1beta1().ReplicaSets(namespace).List(metav1.ListOptions{LabelSelector: labelSelector})
-
-	if err != nil {
-		logrus.Error("Could not get replica sets! ", err)
-	}
-
-	for _, replicaSet := range replicaSets.Items {
-		err := k.Kclient.ExtensionsV1beta1().ReplicaSets(namespace).Delete(replicaSet.Name, &metav1.DeleteOptions{})
-
-		if err != nil {
-			logrus.Errorf("Could not delete replica sets: %s ", replicaSet.Name)
-		} else {
-			logrus.Infof("Deleted replica set: %s", replicaSet.Name)
-		}
-	}
-
-	return nil
-}
-
 // DeleteStatefulSet deletes the data statefulset
 func (k *K8sutil) DeleteStatefulSet(clusterName, namespace string) error {
 
-	// Get list of deployments
+	// Get list of data type statefulsets
 	statefulsets, err := k.Kclient.AppsV1beta1().StatefulSets(namespace).List(metav1.ListOptions{LabelSelector: "component=elasticsearch" + "-" + clusterName + ",role=data"})
 
 	if err != nil {
@@ -552,7 +301,7 @@ func (k *K8sutil) DeleteStatefulSet(clusterName, namespace string) error {
 	}
 
 	for _, statefulset := range statefulsets.Items {
-		//Scale the deployment down to zero (https://github.com/kubernetes/client-go/issues/91)
+		//Scale the statefulset down to zero (https://github.com/kubernetes/client-go/issues/91)
 		statefulset.Spec.Replicas = new(int32)
 		statefulset, err := k.Kclient.AppsV1beta1().StatefulSets(namespace).Update(&statefulset)
 
@@ -573,200 +322,6 @@ func (k *K8sutil) DeleteStatefulSet(clusterName, namespace string) error {
 			logrus.Errorf("Could not delete statefulset: %s ", statefulset.Name)
 		} else {
 			logrus.Infof("Deleted statefulset: %s", statefulset.Name)
-		}
-	}
-
-	return nil
-}
-
-// CreateClientMasterDeployment creates the client or master deployment
-func (k *K8sutil) CreateClientMasterDeployment(deploymentType, baseImage string, replicas *int32, javaOptions string,
-	resources myspec.Resources, imagePullSecrets []myspec.ImagePullSecrets, clusterName, statsdEndpoint, networkHost, namespace string) error {
-
-	component := fmt.Sprintf("elasticsearch-%s", clusterName)
-	discoveryServiceNameCluster := fmt.Sprintf("%s-%s", discoveryServiceName, clusterName)
-
-	var deploymentName, role, isNodeMaster, httpEnable string
-
-	if deploymentType == "client" {
-		httpEnable = "true"
-		deploymentName = clientDeploymentName + "-" + clusterName
-		isNodeMaster = "false"
-		role = "client"
-	} else if deploymentType == "master" {
-		httpEnable = "false"
-		deploymentName = masterDeploymentName + "-" + clusterName
-		isNodeMaster = "true"
-		role = "master"
-	}
-
-	// Check if deployment exists
-	deployment, err := k.Kclient.ExtensionsV1beta1().Deployments(namespace).Get(deploymentName, metav1.GetOptions{})
-
-	if len(deployment.Name) == 0 {
-		logrus.Infof("%s not found, creating...", deploymentName)
-
-		// Parse CPU / Memory
-		limitCPU, _ := resource.ParseQuantity(resources.Limits.CPU)
-		limitMemory, _ := resource.ParseQuantity(resources.Limits.Memory)
-		requestCPU, _ := resource.ParseQuantity(resources.Requests.CPU)
-		requestMemory, _ := resource.ParseQuantity(resources.Requests.Memory)
-
-		deployment := &v1beta1.Deployment{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: deploymentName,
-				Labels: map[string]string{
-					"component": component,
-					"role":      role,
-					"name":      deploymentName,
-				},
-			},
-			Spec: v1beta1.DeploymentSpec{
-				Replicas: replicas,
-				Template: v1.PodTemplateSpec{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"component": component,
-							"role":      role,
-							"name":      deploymentName,
-						},
-						Annotations: map[string]string{
-							"pod.beta.kubernetes.io/init-containers": "[ { \"name\": \"sysctl\", \"image\": \"busybox\", \"imagePullPolicy\": \"IfNotPresent\", \"command\": [\"sysctl\", \"-w\", \"vm.max_map_count=262144\"], \"securityContext\": { \"privileged\": true } }]",
-						},
-					},
-					Spec: v1.PodSpec{
-						Containers: []v1.Container{
-							v1.Container{
-								Name: deploymentName,
-								SecurityContext: &v1.SecurityContext{
-									Privileged: &[]bool{true}[0],
-									Capabilities: &v1.Capabilities{
-										Add: []v1.Capability{
-											"IPC_LOCK",
-										},
-									},
-								},
-								Image:           baseImage,
-								ImagePullPolicy: "Always",
-								Env: []v1.EnvVar{
-									v1.EnvVar{
-										Name: "NAMESPACE",
-										ValueFrom: &v1.EnvVarSource{
-											FieldRef: &v1.ObjectFieldSelector{
-												FieldPath: "metadata.namespace",
-											},
-										},
-									},
-									v1.EnvVar{
-										Name:  "CLUSTER_NAME",
-										Value: "myesdb",
-									},
-									v1.EnvVar{
-										Name:  "NODE_MASTER",
-										Value: isNodeMaster,
-									},
-									v1.EnvVar{
-										Name:  "NODE_DATA",
-										Value: "false",
-									},
-									v1.EnvVar{
-										Name:  "HTTP_ENABLE",
-										Value: httpEnable,
-									},
-									v1.EnvVar{
-										Name:  "ES_JAVA_OPTS",
-										Value: javaOptions,
-									},
-									v1.EnvVar{
-										Name:  "STATSD_HOST",
-										Value: statsdEndpoint,
-									},
-									v1.EnvVar{
-										Name:  "DISCOVERY_SERVICE",
-										Value: discoveryServiceNameCluster,
-									},
-									v1.EnvVar{
-										Name:  "NETWORK_HOST",
-										Value: networkHost,
-									},
-								},
-								Ports: []v1.ContainerPort{
-									v1.ContainerPort{
-										Name:          "transport",
-										ContainerPort: 9300,
-										Protocol:      v1.ProtocolTCP,
-									},
-									v1.ContainerPort{
-										Name:          "http",
-										ContainerPort: 9200,
-										Protocol:      v1.ProtocolTCP,
-									},
-								},
-								VolumeMounts: []v1.VolumeMount{
-									v1.VolumeMount{
-										Name:      "storage",
-										MountPath: "/data",
-									},
-									v1.VolumeMount{
-										Name:      fmt.Sprintf("%s-%s", secretName, clusterName),
-										MountPath: "/elasticsearch/config/certs",
-									},
-								},
-								Resources: v1.ResourceRequirements{
-									Limits: v1.ResourceList{
-										"cpu":    limitCPU,
-										"memory": limitMemory,
-									},
-									Requests: v1.ResourceList{
-										"cpu":    requestCPU,
-										"memory": requestMemory,
-									},
-								},
-							},
-						},
-						Volumes: []v1.Volume{
-							v1.Volume{
-								Name: "storage",
-								VolumeSource: v1.VolumeSource{
-									EmptyDir: &v1.EmptyDirVolumeSource{},
-								},
-							},
-							v1.Volume{
-								Name: fmt.Sprintf("%s-%s", secretName, clusterName),
-								VolumeSource: v1.VolumeSource{
-									Secret: &v1.SecretVolumeSource{
-										SecretName: fmt.Sprintf("%s-%s", secretName, clusterName),
-									},
-								},
-							},
-						},
-						ImagePullSecrets: TemplateImagePullSecrets(imagePullSecrets),
-					},
-				},
-			},
-		}
-
-		_, err := k.Kclient.ExtensionsV1beta1().Deployments(namespace).Create(deployment)
-
-		if err != nil {
-			logrus.Error("Could not create client deployment: ", err)
-			return err
-		}
-	} else {
-		if err != nil {
-			logrus.Error("Could not get client deployment! ", err)
-			return err
-		}
-
-		//scale replicas?
-		if deployment.Spec.Replicas != replicas {
-			deployment.Spec.Replicas = replicas
-
-			_, err := k.Kclient.ExtensionsV1beta1().Deployments(namespace).Update(deployment)
-
-			if err != nil {
-				logrus.Error("Could not scale deployment: ", err)
-			}
 		}
 	}
 
@@ -843,7 +398,6 @@ func (k *K8sutil) CreateDataNodeDeployment(replicas *int32, baseImage, storageCl
 														Key:      "role",
 														Operator: metav1.LabelSelectorOpIn,
 														Values:   []string{"data"},
-
 													},
 												},
 											},
@@ -998,103 +552,4 @@ func (k *K8sutil) CreateDataNodeDeployment(replicas *int32, baseImage, storageCl
 	}
 
 	return nil
-}
-
-// CreateStorageClass creates a storage class
-// NOTE: Right now only creating AWS EBS volumes type gp2
-func (k *K8sutil) CreateStorageClass(zone, storageClassProvisioner, storageType string, clusterName string) error {
-
-	component := "elasticsearch" + "-" + clusterName
-	// Check if storage class exists
-	storageClass, err := k.Kclient.StorageV1beta1().StorageClasses().Get(zone, metav1.GetOptions{})
-
-	if len(storageClass.Name) == 0 {
-		logrus.Infof("StorgeClass %s not found, creating...", zone)
-
-		class := &storage.StorageClass{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: zone,
-				Labels: map[string]string{
-					"component": component,
-				},
-			},
-			Provisioner: storageClassProvisioner,
-			Parameters: map[string]string{
-				"type": storageType,
-			},
-		}
-
-		if zone != "es-default" {
-			class.Parameters["zone"] = zone
-		}
-
-		_, err := k.Kclient.StorageV1beta1().StorageClasses().Create(class)
-
-		if err != nil {
-			logrus.Error("Could not create storage class: ", err)
-			return err
-		}
-	} else if err != nil {
-		logrus.Error("Could not get storage class! ", err)
-		return err
-	}
-
-	return nil
-}
-
-// DeleteStorageClasses removes storage classes tied to the operator
-func (k *K8sutil) DeleteStorageClasses(clusterName string) error {
-	component := "elasticsearch" + "-" + clusterName
-	err := k.Kclient.StorageV1beta1().StorageClasses().DeleteCollection(&metav1.DeleteOptions{}, metav1.ListOptions{LabelSelector: component})
-
-	if err != nil {
-		logrus.Error("Could not delete storageclasses: ", err)
-	} else {
-		logrus.Info("Deleted storageclasses")
-	}
-
-	return nil
-}
-
-// UpdateVolumeReclaimPolicy updates the policy of the volume after it's created:
-// See: https://github.com/kubernetes/kubernetes/issues/38192
-func (k *K8sutil) UpdateVolumeReclaimPolicy(policy, namespace string) {
-
-	var policyType v1.PersistentVolumeReclaimPolicy
-
-	switch policy {
-	case "Delete":
-		policyType = v1.PersistentVolumeReclaimDelete
-		break
-	case "Retain":
-		policyType = v1.PersistentVolumeReclaimRetain
-		break
-	}
-
-	pvc, err := k.Kclient.CoreV1().PersistentVolumeClaims(namespace).List(metav1.ListOptions{
-		LabelSelector: "role=data",
-	})
-
-	if err != nil {
-		return
-	}
-
-	for _, v := range pvc.Items {
-		pv, err := k.Kclient.CoreV1().PersistentVolumes().Get(v.Spec.VolumeName, metav1.GetOptions{})
-
-		if err != nil {
-			continue
-		}
-
-		// Set the policy
-		pv.Spec.PersistentVolumeReclaimPolicy = policyType
-
-		_, err = k.Kclient.CoreV1().PersistentVolumes().Update(pv)
-
-		if err != nil {
-			logrus.Error("Could not update pv! ", err)
-			continue
-		}
-
-	}
 }
